Added design for LinkedIn Connections panel test
We know very few people use the LinkedIn Connections feature but we're not sure if it's because it's too easy to overlook or because they're really not that interested in it. This test is designed to amplify its visibility and combine two different but related panels into one near the top. Once the test is complete it should be reverted or pushed forward to a better design that better dials in its visibility alongside other page elements.