Add Sign In Test to Stack Selection Flow
We want to see if we can improve the UX of the wizard stack selection flow by introducing a sign in push at the beginning as an alternative to the "Role Selection" screen. By leveraging Position (and eventually Company) data from a LinkedIn connection, we should be able to serve up better categories and lists of products in the stack selection screen.