After-merge changes latently appearing in Kactus
In Kactus there are 1264 changed files appearing after the previous merge and after syncing with the origin branch. I did resolve three conflicts in GitHub before merging that had to do with the _archive field values and I resolved them in favor of the pushed branch thinking this would cause fewer changes locally. Appears that's not the case, or something else is going on. Going to commit these changes and attempt a merge before switching branches again. If we can't come to an understanding or an improvement in the way Kactus is tracking/changing files, this is going to be a consistent pain point for our design team. My best guess at present is that it has to do with when Sketch opens or closes and changes that are written at those points.